+++ database/migrations/2022_06_14_000000_version_602.php
@@ -1,0 +1,29 @@
+<?php
+
+use BT\Modules\Settings\Models\Setting;
+use Illuminate\Database\Migrations\Migration;
+
+class Version602 extends Migration
+{
+
+    /**
+     * Run the migrations.
+     * @table payments_custom
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Setting::saveByKey('version', '6.0.2');
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+     public function down()
+     {
+         Setting::saveByKey('version', '6.0.1');
+     }
+}